增加增量同步用户信息相关接口访问

+++ app/src/main/java/net/ischool/isus/model/Uids.kt
@@ -3,10 +3,10 @@ package net.ischool.isus.model
 import com.google.gson.annotations.SerializedName
 
 /**
- * Description
+ * 用户ID集合
  *
  * Author: Walker
  * Email: zhaocework@gmail.com
  * Date: 2019-05-30
  */
-data class Uids(@SerializedName(value = "list") val uids: List<Long>)
+data class Uids(@SerializedName(value = "list") val uids: List<Int>)
